CacheStats: use long/double instead of int/float

+++ projects/bdd/src/main/java/net/sf/javabdd/BDDFactory.java
@@ -707,15 +707,15 @@ public abstract class BDDFactory {
    * @version $Id: BDDFactory.java,v 1.18 2005/10/12 10:27:08 joewhaley Exp $
    */
   public static class CacheStats {
-    public int uniqueAccess;
-    public int uniqueChain;
-    public int uniqueHit;
-    public int uniqueMiss;
-    public int uniqueTrivial;
-    public int opHit;
-    public int opMiss;
-    public int opOverwrite;
-    public int swapCount;
+    public long uniqueAccess;
+    public long uniqueChain;
+    public long uniqueHit;
+    public long uniqueMiss;
+    public long uniqueTrivial;
+    public long opHit;
+    public long opMiss;
+    public long opOverwrite;
+    public long swapCount;
 
     protected CacheStats() {}
 
@@ -758,9 +758,9 @@ public abstract class BDDFactory {
       sb.append(newLine);
       sb.append("=> Hit rate =   ");
       if (uniqueHit + uniqueMiss > 0) {
-        sb.append(((float) uniqueHit) / ((float) uniqueHit + uniqueMiss));
+        sb.append(((double) uniqueHit) / ((double) uniqueHit + uniqueMiss));
       } else {
-        sb.append((float) 0);
+        sb.append((double) 0);
       }
       sb.append(newLine);
       sb.append("Operator Hits:  ");
@@ -774,9 +774,9 @@ public abstract class BDDFactory {
       sb.append(newLine);
       sb.append("=> Hit rate =   ");
       if (opHit + opMiss > 0) {
-        sb.append(((float) opHit) / ((float) opHit + opMiss));
+        sb.append(((double) opHit) / ((double) opHit + opMiss));
       } else {
-        sb.append((float) 0);
+        sb.append((double) 0);
       }
       sb.append(newLine);
       sb.append("Swap count =    ");
